move config definition into a single xlayer file
